+++ src/components/openapi/Introduction.tsx
@@ -40,9 +40,7 @@ export function Introduction() {
         </div>
       )}
 
-      {info.description && (
-        <Markdown content={info.description} class="max-w-none mt-8" />
-      )}
+      {info.description && <Markdown content={info.description} class="max-w-none mt-8" />}
     </div>
   );
 }